+++ conjur/user/__init__.py
@@ -1,0 +1,9 @@
+"""
+User module
+
+This metafile includes all the functionality that will be exposed
+when you install this module
+"""
+from conjur.user.user_controller import UserController
+from conjur.user.user_logic import UserLogic
+from conjur.user.user_resource_data import UserResourceData