+++ test/test_unit_api.py
@@ -214,7 +214,6 @@ class ApiTest(unittest.TestCase):
                               query={},
                               ssl_verify=True)
 
-    test_get_variable_invokes_http_client_correctly.tester=True
     @patch('conjur.api.invoke_endpoint', return_value=MockClientResponse())
     def test_get_variable_with_version_invokes_http_client_correctly(self, mock_http_client):
         api = Api(url='http://localhost', login_id='mylogin', api_key='apikey')